+++ x/auth/simulation/decoder.go
@@ -7,6 +7,7 @@ import (
 	gogotypes "github.com/gogo/protobuf/types"
 
 	"github.com/cosmos/cosmos-sdk/codec"
+	sdktypes "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/kv"
 	"github.com/cosmos/cosmos-sdk/x/auth/types"
 )
@@ -40,6 +41,20 @@ func NewDecodeStore(ak AuthUnmarshaler) func(kvA, kvB kv.Pair) string {
 			ak.GetCodec().MustUnmarshal(kvB.Value, &globalAccNumberB)
 
 			return fmt.Sprintf("GlobalAccNumberA: %d\nGlobalAccNumberB: %d", globalAccNumberA, globalAccNumberB)
+
+		case bytes.Equal(kvA.Key[:len(types.AccountNumberStoreKeyPrefix)], types.AccountNumberStoreKeyPrefix):
+			var accNumA, accNumB sdktypes.AccAddress
+			err := accNumA.Unmarshal(kvA.Value)
+			if err != nil {
+				panic(err)
+			}
+
+			err = accNumB.Unmarshal(kvB.Value)
+			if err != nil {
+				panic(err)
+			}
+
+			return fmt.Sprintf("AccNumA: %s\nAccNumB: %s", accNumA, accNumB)
 
 		default:
 			panic(fmt.Sprintf("unexpected %s key %X (%s)", types.ModuleName, kvA.Key, kvA.Key))