+++ x/auth/types/genesis.go
@@ -75,6 +75,43 @@ func ValidateGenesis(data GenesisState) error {
 
 // SanitizeGenesisAccounts sorts accounts and coin sets.
 func SanitizeGenesisAccounts(genAccs GenesisAccounts) GenesisAccounts {
+	// Make sure there aren't any duplicated account numbers by fixing the duplicates with the lowest unused values.
+	// seenAccNum = easy lookup for used account numbers.
+	seenAccNum := map[uint64]bool{}
+	// dupAccNum = a map of account number to accounts with duplicate account numbers (excluding the 1st one seen).
+	dupAccNum := map[uint64]GenesisAccounts{}
+	for _, acc := range genAccs {
+		num := acc.GetAccountNumber()
+		if !seenAccNum[num] {
+			seenAccNum[num] = true
+		} else {
+			dupAccNum[num] = append(dupAccNum[num], acc)
+		}
+	}
+	// dupAccNums a sorted list of the account numbers with duplicates.
+	var dupAccNums []uint64
+	for num := range dupAccNum {
+		dupAccNums = append(dupAccNums, num)
+	}
+	sort.Slice(dupAccNums, func(i, j int) bool {
+		return dupAccNums[i] < dupAccNums[j]
+	})
+	// Change the account number of the duplicated ones to the first unused value.
+	globalNum := uint64(0)
+	for _, dupNum := range dupAccNums {
+		accs := dupAccNum[dupNum]
+		for _, acc := range accs {
+			for seenAccNum[globalNum] {
+				globalNum++
+			}
+			if err := acc.SetAccountNumber(globalNum); err != nil {
+				panic(err)
+			}
+			seenAccNum[globalNum] = true
+		}
+	}
+
+	// Then sort them all by account number.
 	sort.Slice(genAccs, func(i, j int) bool {
 		return genAccs[i].GetAccountNumber() < genAccs[j].GetAccountNumber()
 	})